+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+# v0.6.1
+
+- Merged in hotfix from parent project: Validator\Financial\Bic updated for ISO 9362.
+- Update `make composer` to install composer without the need of `curl`
+- Improved documentation
+
 # v0.6.0
 
 - Split off the validator as a separate project